Rename CacheTransformer to DocumentTransformer
