+++ lib/orbit-common/cache.js
@@ -6,7 +6,7 @@ import { eq } from 'orbit/lib/eq';
 import { deprecate } from 'orbit/lib/deprecate';
 import CacheIntegrityProcessor from 'orbit-common/operation-processors/cache-integrity-processor';
 import SchemaConsistencyProcessor from 'orbit-common/operation-processors/schema-consistency-processor';
-import CacheTransformer from './cache-transformer';
+import DocumentTransformer from './document-transformer';
 
 /**
  `Cache` provides a thin wrapper over an internally maintained instance of a
@@ -43,7 +43,7 @@ export default Class.extend({
     var _this = this;
     var retrieve = function(path) { return _this.retrieve(path); };
     var transformTarget = function(operation, invert){ return _this._doc.transform(operation, invert); };
-    this._cacheTransformer = new CacheTransformer(retrieve, transformTarget, this._schema, processors);
+    this._documentTransformer = new DocumentTransformer(retrieve, transformTarget, this._schema, processors);
   },
 
 
@@ -57,7 +57,7 @@ export default Class.extend({
   reset: function(data) {
     this._doc.reset(data);
     this._schema.registerAllKeys(data);
-    this._cacheTransformer.reset(data);
+    this._documentTransformer.reset(data);
   },
 
   /**
@@ -134,7 +134,7 @@ export default Class.extend({
    @returns {Array or undefined} Array of inverse operations
    */
   transform: function(operation) {
-    return this._cacheTransformer.transform(this._normalizeOperation(operation));
+    return this._documentTransformer.transform(this._normalizeOperation(operation));
   },
 
   _normalizeOperation: function(operation) {
