+++ lib/orbit-common/document-transformer.js
